+++ packages/button/src/utils/index.android.js
@@ -1,3 +1,1 @@
-import { androidLetterSpacing } from "@times-components/utils";
-
-export default androidLetterSpacing;
+export default title => title.toUpperCase();