+++ packages/context/src/defaults/scales.js
@@ -1,0 +1,5 @@
+export default Object.freeze({
+  medium: "medium",
+  large: "large",
+  xlarge: "xlarge"
+});