+++ packages/icons/__tests__/ios/icons-dim.ios.test.js
@@ -1,5 +1,3 @@
 import shared from "../shared-dim.native";
 
-describe("ios", () => {
-  shared();
-});
+shared();