+++ packages/icons/__tests__/web/icons-dim.web.test.js
@@ -1,5 +1,3 @@
 import shared from "../shared-dim.web";
 
-describe("web", () => {
-  shared();
-});
+shared();