+++ packages/image/__tests__/web/jest.config.js
@@ -2,6 +2,6 @@ const jestConfigurator = require("@times-components/jest-configurator").default;
 const path = require("path");
 
 module.exports = jestConfigurator("web", __dirname, {
-  coverageIgnoreGlobs: ["add-missing-protocol.js"],
+  coverageIgnoreGlobs: ["index.js", "modal-image.web.js"],
   setupTestFrameworkScriptFile: path.join(__dirname, "./serializers")
 });