+++ packages/related-articles/fixtures/standard/3-articles.js
@@ -373,38 +373,18 @@ const defaultSecondUrl =
   "https://www.thetimes.co.uk/article/britain-to-get-bayeux-tapestry-as-macron-agrees-loan-n5brflnjx";
 
 const defaultThirdCrop169 =
-  "https://www.thetimes.co.uk/imageserver/image/methode%2Ftimes%2Fprod%2Fweb%2Fbin%2F82723c10-fb7f-11e7-a987-7fcf5e9983dc.jpg?crop=4886%2C2748%2C92%2C108";
-const defaultThirdHeadline = "Bayeux Tapestry to be displayed in Britain";
+  "https://www.thetimes.co.uk/imageserver/image/%2Fmethode%2Ftimes%2Fprod%2Fweb%2Fbin%2Ffa613a54-39c4-11e8-b5b4-b935584040f4.jpg?crop=939%2C528%2C0%2C0";
+const defaultThirdHeadline =
+  "YouTube ignored police pleas to remove threatening videos";
 const defaultThirdSummary105 = [
   {
     name: "paragraph",
-    attributes: {},
     children: [
       {
         name: "text",
         attributes: {
           value:
-            "The first loan of the Bayeux Tapestry outside France for 950 years, revealed today by "
-        },
-        children: []
-      },
-      {
-        name: "italic",
-        attributes: {},
-        children: [
-          {
-            name: "text",
-            attributes: {
-              value: "The Times"
-            },
-            children: []
-          }
-        ]
-      },
-      {
-        name: "text",
-        attributes: {
-          value: ", is"
+            "Light gleamed from the blade of what appeared to be a combat knife in a video of a huddled group of"
         },
         children: []
       }
@@ -414,33 +394,12 @@ const defaultThirdSummary105 = [
 const defaultThirdSummary125 = [
   {
     name: "paragraph",
-    attributes: {},
     children: [
       {
         name: "text",
         attributes: {
           value:
-            "The first loan of the Bayeux Tapestry outside France for 950 years, revealed today by "
-        },
-        children: []
-      },
-      {
-        name: "italic",
-        attributes: {},
-        children: [
-          {
-            name: "text",
-            attributes: {
-              value: "The Times"
-            },
-            children: []
-          }
-        ]
-      },
-      {
-        name: "text",
-        attributes: {
-          value: ", is expected to be to the"
+            "Light gleamed from the blade of what appeared to be a combat knife in a video of a huddled group of teenage boys in hoods and"
         },
         children: []
       }
@@ -450,87 +409,27 @@ const defaultThirdSummary125 = [
 const defaultThirdSummary145 = [
   {
     name: "paragraph",
-    attributes: {},
     children: [
       {
         name: "text",
         attributes: {
           value:
-            "The first loan of the Bayeux Tapestry outside France for 950 years, revealed today by "
-        },
-        children: []
-      },
-      {
-        name: "italic",
-        attributes: {},
-        children: [
-          {
-            name: "text",
-            attributes: {
-              value: "The Times"
-            },
-            children: []
-          }
-        ]
-      },
-      {
-        name: "text",
-        attributes: {
-          value: ", is expected to be to the British Museum in 2022"
+            "Light gleamed from the blade of what appeared to be a combat knife in a video of a huddled group of teenage boys in hoods and masks in Haringey"
         },
         children: []
       }
     ]
-  },
-  {
-    name: "paragraph",
-    attributes: {},
-    children: []
   }
 ];
 const defaultThirdSummary160 = [
   {
     name: "paragraph",
-    attributes: {},
     children: [
       {
         name: "text",
         attributes: {
           value:
-            "The first loan of the Bayeux Tapestry outside France for 950 years, revealed today by "
-        },
-        children: []
-      },
-      {
-        name: "italic",
-        attributes: {},
-        children: [
-          {
-            name: "text",
-            attributes: {
-              value: "The Times"
-            },
-            children: []
-          }
-        ]
-      },
-      {
-        name: "text",
-        attributes: {
-          value: ", is expected to be to the British Museum in 2022."
-        },
-        children: []
-      }
-    ]
-  },
-  {
-    name: "paragraph",
-    attributes: {},
-    children: [
-      {
-        name: "text",
-        attributes: {
-          value: "Although the"
+            "Light gleamed from the blade of what appeared to be a combat knife in a video of a huddled group of teenage boys in hoods and masks in Haringey, north London"
         },
         children: []
       }
@@ -540,46 +439,12 @@ const defaultThirdSummary160 = [
 const defaultThirdSummary175 = [
   {
     name: "paragraph",
-    attributes: {},
     children: [
       {
         name: "text",
         attributes: {
           value:
-            "The first loan of the Bayeux Tapestry outside France for 950 years, revealed today by "
-        },
-        children: []
-      },
-      {
-        name: "italic",
-        attributes: {},
-        children: [
-          {
-            name: "text",
-            attributes: {
-              value: "The Times"
-            },
-            children: []
-          }
-        ]
-      },
-      {
-        name: "text",
-        attributes: {
-          value: ", is expected to be to the British Museum in 2022."
-        },
-        children: []
-      }
-    ]
-  },
-  {
-    name: "paragraph",
-    attributes: {},
-    children: [
-      {
-        name: "text",
-        attributes: {
-          value: "Although the museum stopped"
+            "Light gleamed from the blade of what appeared to be a combat knife in a video of a huddled group of teenage boys in hoods and masks in Haringey, north London. Within a week a"
         },
         children: []
       }
@@ -589,57 +454,21 @@ const defaultThirdSummary175 = [
 const defaultThirdSummary225 = [
   {
     name: "paragraph",
-    attributes: {},
     children: [
       {
         name: "text",
         attributes: {
           value:
-            "The first loan of the Bayeux Tapestry outside France for 950 years, revealed today by "
-        },
-        children: []
-      },
-      {
-        name: "italic",
-        attributes: {},
-        children: [
-          {
-            name: "text",
-            attributes: {
-              value: "The Times"
-            },
-            children: []
-          }
-        ]
-      },
-      {
-        name: "text",
-        attributes: {
-          value: ", is expected to be to the British Museum in 2022."
-        },
-        children: []
-      }
-    ]
-  },
-  {
-    name: "paragraph",
-    attributes: {},
-    children: [
-      {
-        name: "text",
-        attributes: {
-          value:
-            "Although the museum stopped short of confirming that it had secured the loan"
+            "Light gleamed from the blade of what appeared to be a combat knife in a video of a huddled group of teenage boys in hoods and masks in Haringey, north London. Within a week a boy linked to a group known as G Lanes had been"
         },
         children: []
       }
     ]
   }
 ];
-const defaultThirdTitle =
-  "Bayeux Tapestry 1067: Battle of Hastings, 14 October 1066. The death of Harold II, last Anglo-Saxon king of England. Left, figure pulling arrow from eye and then being cut down by Norman knight. Armour Chain Mail Sword Axe Textile";
+const defaultThirdTitle = "Some scary people";
 const defaultThirdUrl =
-  "https://www.thetimes.co.uk/article/britain-to-get-bayeux-tapestry-as-macron-agrees-loan-n5brflnjx";
+  "https://www.thetimes.co.uk/article/youtube-ignored-police-pleas-to-remove-threatening-videos-62v2mqp23";
 
 export default (
   {
@@ -755,9 +584,11 @@ export default (
         id: "8b47c1b8-fb05-11e7-a987-7fcf5e99898c",
         publishedTime: "2018-01-17T12:00:00.000Z",
         leadAsset: {
-          title: thirdTitle,
-          crop169: {
-            url: thirdCrop169
+          posterImage: {
+            title: thirdTitle,
+            crop169: {
+              url: thirdCrop169
+            }
           }
         },
         label: "BAYEUX TAPESTRY",