+++ ios-app/CHANGELOG.md
@@ -3,6 +3,78 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.2.20"></a>
+## [0.2.20](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.2.19...@times-components/ios-app@0.2.20) (2018-08-01)
+
+
+
+
+**Note:** Version bump only for package @times-components/ios-app
+
+<a name="0.2.19"></a>
+## [0.2.19](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.2.18...@times-components/ios-app@0.2.19) (2018-08-01)
+
+
+
+
+**Note:** Version bump only for package @times-components/ios-app
+
+<a name="0.2.18"></a>
+## [0.2.18](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.2.17...@times-components/ios-app@0.2.18) (2018-07-31)
+
+
+
+
+**Note:** Version bump only for package @times-components/ios-app
+
+<a name="0.2.17"></a>
+## [0.2.17](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.2.16...@times-components/ios-app@0.2.17) (2018-07-31)
+
+
+
+
+**Note:** Version bump only for package @times-components/ios-app
+
+<a name="0.2.16"></a>
+## [0.2.16](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.2.15...@times-components/ios-app@0.2.16) (2018-07-31)
+
+
+
+
+**Note:** Version bump only for package @times-components/ios-app
+
+<a name="0.2.15"></a>
+## [0.2.15](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.2.14...@times-components/ios-app@0.2.15) (2018-07-31)
+
+
+
+
+**Note:** Version bump only for package @times-components/ios-app
+
+<a name="0.2.14"></a>
+## [0.2.14](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.2.13...@times-components/ios-app@0.2.14) (2018-07-31)
+
+
+
+
+**Note:** Version bump only for package @times-components/ios-app
+
+<a name="0.2.13"></a>
+## [0.2.13](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.2.12...@times-components/ios-app@0.2.13) (2018-07-26)
+
+
+
+
+**Note:** Version bump only for package @times-components/ios-app
+
+<a name="0.2.12"></a>
+## [0.2.12](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.2.11...@times-components/ios-app@0.2.12) (2018-07-26)
+
+
+
+
+**Note:** Version bump only for package @times-components/ios-app
+
 <a name="0.2.11"></a>
 ## [0.2.11](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.2.10...@times-components/ios-app@0.2.11) (2018-07-26)