+++ packages/link/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.0.10"></a>
+## [2.0.10](https://github.com/newsuk/times-components/compare/@times-components/link@2.0.9...@times-components/link@2.0.10) (2018-07-31)
+
+
+
+
+**Note:** Version bump only for package @times-components/link
+
+<a name="2.0.9"></a>
+## [2.0.9](https://github.com/newsuk/times-components/compare/@times-components/link@2.0.8...@times-components/link@2.0.9) (2018-07-31)
+
+
+
+
+**Note:** Version bump only for package @times-components/link
+
+<a name="2.0.8"></a>
+## [2.0.8](https://github.com/newsuk/times-components/compare/@times-components/link@2.0.7...@times-components/link@2.0.8) (2018-07-26)
+
+
+
+
+**Note:** Version bump only for package @times-components/link
+
+<a name="2.0.7"></a>
+## [2.0.7](https://github.com/newsuk/times-components/compare/@times-components/link@2.0.6...@times-components/link@2.0.7) (2018-07-26)
+
+
+
+
+**Note:** Version bump only for package @times-components/link
+
 <a name="2.0.6"></a>
 ## [2.0.6](https://github.com/newsuk/times-components/compare/@times-components/link@2.0.5...@times-components/link@2.0.6) (2018-07-25)