+++ packages/pages/CHANGELOG.md
@@ -3,6 +3,78 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.2.11"></a>
+## [0.2.11](https://github.com/newsuk/times-components/compare/@times-components/pages@0.2.10...@times-components/pages@0.2.11) (2018-08-01)
+
+
+
+
+**Note:** Version bump only for package @times-components/pages
+
+<a name="0.2.10"></a>
+## [0.2.10](https://github.com/newsuk/times-components/compare/@times-components/pages@0.2.9...@times-components/pages@0.2.10) (2018-08-01)
+
+
+
+
+**Note:** Version bump only for package @times-components/pages
+
+<a name="0.2.9"></a>
+## [0.2.9](https://github.com/newsuk/times-components/compare/@times-components/pages@0.2.8...@times-components/pages@0.2.9) (2018-07-31)
+
+
+
+
+**Note:** Version bump only for package @times-components/pages
+
+<a name="0.2.8"></a>
+## [0.2.8](https://github.com/newsuk/times-components/compare/@times-components/pages@0.2.7...@times-components/pages@0.2.8) (2018-07-31)
+
+
+
+
+**Note:** Version bump only for package @times-components/pages
+
+<a name="0.2.7"></a>
+## [0.2.7](https://github.com/newsuk/times-components/compare/@times-components/pages@0.2.6...@times-components/pages@0.2.7) (2018-07-31)
+
+
+
+
+**Note:** Version bump only for package @times-components/pages
+
+<a name="0.2.6"></a>
+## [0.2.6](https://github.com/newsuk/times-components/compare/@times-components/pages@0.2.5...@times-components/pages@0.2.6) (2018-07-31)
+
+
+
+
+**Note:** Version bump only for package @times-components/pages
+
+<a name="0.2.5"></a>
+## [0.2.5](https://github.com/newsuk/times-components/compare/@times-components/pages@0.2.4...@times-components/pages@0.2.5) (2018-07-31)
+
+
+
+
+**Note:** Version bump only for package @times-components/pages
+
+<a name="0.2.4"></a>
+## [0.2.4](https://github.com/newsuk/times-components/compare/@times-components/pages@0.2.3...@times-components/pages@0.2.4) (2018-07-26)
+
+
+
+
+**Note:** Version bump only for package @times-components/pages
+
+<a name="0.2.3"></a>
+## [0.2.3](https://github.com/newsuk/times-components/compare/@times-components/pages@0.2.2...@times-components/pages@0.2.3) (2018-07-26)
+
+
+
+
+**Note:** Version bump only for package @times-components/pages
+
 <a name="0.2.2"></a>
 ## [0.2.2](https://github.com/newsuk/times-components/compare/@times-components/pages@0.2.1...@times-components/pages@0.2.2) (2018-07-26)