+++ packages/provider-test-tools/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.15.5"></a>
+## [0.15.5](https://github.com/newsuk/times-components/compare/@times-components/provider-test-tools@0.15.4...@times-components/provider-test-tools@0.15.5) (2018-07-31)
+
+
+
+
+**Note:** Version bump only for package @times-components/provider-test-tools
+
+<a name="0.15.4"></a>
+## [0.15.4](https://github.com/newsuk/times-components/compare/@times-components/provider-test-tools@0.15.3...@times-components/provider-test-tools@0.15.4) (2018-07-31)
+
+
+
+
+**Note:** Version bump only for package @times-components/provider-test-tools
+
+<a name="0.15.3"></a>
+## [0.15.3](https://github.com/newsuk/times-components/compare/@times-components/provider-test-tools@0.15.2...@times-components/provider-test-tools@0.15.3) (2018-07-26)
+
+
+### Bug Fixes
+
+* Allow poster image to show with lead asset videos ([#1145](https://github.com/newsuk/times-components/issues/1145)) ([9bf5a13](https://github.com/newsuk/times-components/commit/9bf5a13))
+
+
+
+
+<a name="0.15.2"></a>
+## [0.15.2](https://github.com/newsuk/times-components/compare/@times-components/provider-test-tools@0.15.1...@times-components/provider-test-tools@0.15.2) (2018-07-26)
+
+
+
+
+**Note:** Version bump only for package @times-components/provider-test-tools
+
 <a name="0.15.1"></a>
 ## [0.15.1](https://github.com/newsuk/times-components/compare/@times-components/provider-test-tools@0.15.0...@times-components/provider-test-tools@0.15.1) (2018-07-25)