+++ packages/tealium-utils/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.5.8"></a>
+## [0.5.8](https://github.com/newsuk/times-components/compare/@times-components/tealium-utils@0.5.7...@times-components/tealium-utils@0.5.8) (2018-07-31)
+
+
+
+
+**Note:** Version bump only for package @times-components/tealium-utils
+
+<a name="0.5.7"></a>
+## [0.5.7](https://github.com/newsuk/times-components/compare/@times-components/tealium-utils@0.5.6...@times-components/tealium-utils@0.5.7) (2018-07-31)
+
+
+
+
+**Note:** Version bump only for package @times-components/tealium-utils
+
+<a name="0.5.6"></a>
+## [0.5.6](https://github.com/newsuk/times-components/compare/@times-components/tealium-utils@0.5.5...@times-components/tealium-utils@0.5.6) (2018-07-26)
+
+
+
+
+**Note:** Version bump only for package @times-components/tealium-utils
+
+<a name="0.5.5"></a>
+## [0.5.5](https://github.com/newsuk/times-components/compare/@times-components/tealium-utils@0.5.4...@times-components/tealium-utils@0.5.5) (2018-07-26)
+
+
+
+
+**Note:** Version bump only for package @times-components/tealium-utils
+
 <a name="0.5.4"></a>
 ## [0.5.4](https://github.com/newsuk/times-components/compare/@times-components/tealium-utils@0.5.3...@times-components/tealium-utils@0.5.4) (2018-07-25)