+++ packages/tracking/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.0.8"></a>
+## [2.0.8](https://github.com/newsuk/times-components/compare/@times-components/tracking@2.0.7...@times-components/tracking@2.0.8) (2018-07-31)
+
+
+
+
+**Note:** Version bump only for package @times-components/tracking
+
+<a name="2.0.7"></a>
+## [2.0.7](https://github.com/newsuk/times-components/compare/@times-components/tracking@2.0.6...@times-components/tracking@2.0.7) (2018-07-31)
+
+
+
+
+**Note:** Version bump only for package @times-components/tracking
+
+<a name="2.0.6"></a>
+## [2.0.6](https://github.com/newsuk/times-components/compare/@times-components/tracking@2.0.5...@times-components/tracking@2.0.6) (2018-07-26)
+
+
+
+
+**Note:** Version bump only for package @times-components/tracking
+
+<a name="2.0.5"></a>
+## [2.0.5](https://github.com/newsuk/times-components/compare/@times-components/tracking@2.0.4...@times-components/tracking@2.0.5) (2018-07-26)
+
+
+
+
+**Note:** Version bump only for package @times-components/tracking
+
 <a name="2.0.4"></a>
 ## [2.0.4](https://github.com/newsuk/times-components/compare/@times-components/tracking@2.0.3...@times-components/tracking@2.0.4) (2018-07-25)